Major Core refactoring
A new abstract class `AbstractPage` was created to resolve method conflicts between `Page` and `Site`.
